[refactor] Split the leave protocol and convergence tick out
Two sections holding their own state with no shared callers: leaving a
space with its two replay lanes, and the level-triggered re-drive that an
app restart used to be. The announce ledger joins the shared registries,
having four writers outside the tick.
